+++ versoes/testes/db/re/sql/improvise_db.sql
@@ -3,8 +3,8 @@
 -- 
 
 CREATE TABLE IF NOT EXISTS `improvise_db` (
-	`SkillID` int(11) unsigned NOT NULL default '0',
-	`Rate` mediumint(9) unsigned NOT NULL default '0',
+	`SkillID` smallint(2) unsigned NOT NULL default '0',
+	`Rate` smallint(2) unsigned NOT NULL default '0',
 	PRIMARY KEY (`SkillID`)
 ) ENGINE=MyISAM DEFAULT CHARSET=latin1;
