+++ versoes/testes/db/re/sql/item_trade.sql
@@ -3,9 +3,9 @@
 --  
 
 CREATE TABLE IF NOT EXISTS `item_trade` (
-  `ItemID` int(11) unsigned NOT NULL default '0',
-  `TradeMask` smallint(6) unsigned NOT NULL default '0',
-  `GMLvl` smallint(6) unsigned NOT NULL default '0',
+  `ItemID` smallint(2) unsigned NOT NULL default '0',
+  `TradeMask` tinyint(1) unsigned NOT NULL default '0',
+  `GMLvl` tinyint(1) unsigned NOT NULL default '0',
    PRIMARY KEY (`GroupID`)
 ) ENGINE=MyISAM DEFAULT CHARSET=latin1;
