+++ versoes/testes/db/re/sql/reproduce_db.sql
@@ -3,7 +3,7 @@
 -- 
 
 CREATE TABLE IF NOT EXISTS `reproduce_db` (
-	`SkillID` int(11) unsigned NOT NULL default '0',
+	`SkillID` smallint(2) unsigned NOT NULL default '0',
 	PRIMARY KEY (`SkillID`)
 ) ENGINE=MyISAM DEFAULT CHARSET=latin1;
